Move controllers/index.ts -> ControllerFactory.ts
